♻️ [RUMF-1267] split some imports
By splitting imports, we import less stuff, which removes a few cycles
