+++ packages/core/src/tools/createEventRateLimiter.ts
@@ -1,5 +1,7 @@
-import type { RawError } from '..'
-import { clocksNow, ErrorSource, ONE_MINUTE } from '..'
+import type { RawError } from './error'
+import { ErrorSource } from './error'
+import { clocksNow } from './timeUtils'
+import { ONE_MINUTE } from './utils'
 
 export type EventRateLimiter = ReturnType<typeof createEventRateLimiter>
